permisson changed

+++ public/scripts/delete_profile.php
@@ -7,36 +7,41 @@ header('Content-Type: application/json');
 mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT);
 session_start();
 $uid = $_SESSION['uid'] ?? null;
+$utype = $_SESSION['utype'] ?? null;
 
-if (isset ($_POST['uName'])) {
-    $uName = $_POST['uName'];
+if ($utype == 1) {
+    if (isset ($_POST['uName'])) {
+        $uName = $_POST['uName'];
 
-    $sql = "SELECT userId FROM users where uName=?";
-    $stmt = $conn->prepare($sql);
-    $stmt->bind_param("s", $uName);
-    $stmt->execute();
-    $stmt->store_result();
-    if ($stmt->num_rows > 0) {
-        $stmt->bind_result($uid1);
-        $stmt->fetch();
+        $sql = "SELECT userId FROM users where uName=?";
+        $stmt = $conn->prepare($sql);
+        $stmt->bind_param("s", $uName);
+        $stmt->execute();
+        $stmt->store_result();
+        if ($stmt->num_rows > 0) {
+            $stmt->bind_result($uid1);
+            $stmt->fetch();
 
-        $sql1 = "DELETE FROM users WHERE userId=?";
-        $stmt1 = $conn->prepare($sql1);
-        $stmt1->bind_param("i", $uid1);
-        $stmt1->execute();
+            $sql1 = "DELETE FROM users WHERE userId=?";
+            $stmt1 = $conn->prepare($sql1);
+            $stmt1->bind_param("i", $uid1);
+            $stmt1->execute();
 
-        if ($stmt1->execute()) {
-            echo json_encode(['success' => 'User deleted successfully.']);
-            (header('../pages/discussion.php'));
+            if ($stmt1->execute()) {
+                echo json_encode(['success' => 'User deleted successfully.']);
+                (header('../pages/discussion.php'));
+            } else {
+                echo json_encode(['error' => 'User deletion failed.']);
+            }
         } else {
-            echo json_encode(['error' => 'User deletion failed.']);
+            echo json_encode(['error' => 'User not found.']);
         }
     } else {
-        echo json_encode(['error' => 'User not found.']);
+        echo json_encode(['error' => 'uName not provided.']);
+
     }
 } else {
-    echo json_encode(['error' => 'uName not provided.']);
-
+    echo json_encode(['error' => 'NOT AN ADMIN NAUGHTY NAUGHTY.']);
 }
 $conn->close();
